--HG--
branch : soc.2013.gobjectification
rename : libpurple/ciphers/aes.c => libpurple/ciphers/aescipher.c
rename : libpurple/ciphers/des.c => libpurple/ciphers/descipher.c
rename : libpurple/ciphers/hmac.c => libpurple/ciphers/hmaccipher.c
rename : libpurple/ciphers/pbkdf2.c => libpurple/ciphers/pbkdf2cipher.c

+++ libpurple/protocols/jabber/jabber.c
@@ -582,7 +582,7 @@ int jabber_prpl_send_raw(PurpleConnection *gc, const char *buf, int len)
 	 */
 
 	jabber_send_raw(js, buf, len);
-	return (len < 0 ? strlen(buf) : len);
+	return (len < 0 ? (int)strlen(buf) : len);
 }
 
 void jabber_send_signal_cb(PurpleConnection *pc, xmlnode **packet,
