+++ libpurple/protocols/jabber/jutil.c
@@ -676,7 +676,7 @@ static const struct {
 const char *
 jabber_buddy_state_get_name(const JabberBuddyState state)
 {
-	int i;
+	gsize i;
 	for (i = 0; i < G_N_ELEMENTS(jabber_statuses); ++i)
 		if (jabber_statuses[i].state == state)
 			return _(jabber_statuses[i].readable);
@@ -687,7 +687,7 @@ jabber_buddy_state_get_name(const JabberBuddyState state)
 JabberBuddyState
 jabber_buddy_status_id_get_state(const char *id)
 {
-	int i;
+	gsize i;
 	if (!id)
 		return JABBER_BUDDY_STATE_UNKNOWN;
 
@@ -700,7 +700,7 @@ jabber_buddy_status_id_get_state(const char *id)
 
 JabberBuddyState jabber_buddy_show_get_state(const char *id)
 {
-	int i;
+	gsize i;
 
 	g_return_val_if_fail(id != NULL, JABBER_BUDDY_STATE_UNKNOWN);
 
@@ -716,7 +716,7 @@ JabberBuddyState jabber_buddy_show_get_state(const char *id)
 const char *
 jabber_buddy_state_get_show(JabberBuddyState state)
 {
-	int i;
+	gsize i;
 	for (i = 0; i < G_N_ELEMENTS(jabber_statuses); ++i)
 		if (state == jabber_statuses[i].state)
 			return jabber_statuses[i].show;
@@ -727,7 +727,7 @@ jabber_buddy_state_get_show(JabberBuddyState state)
 const char *
 jabber_buddy_state_get_status_id(JabberBuddyState state)
 {
-	int i;
+	gsize i;
 	for (i = 0; i < G_N_ELEMENTS(jabber_statuses); ++i)
 		if (state == jabber_statuses[i].state)
 			return jabber_statuses[i].status_id;
